TabletDebuggerViewModel/Statistics: Fix child collection event signaling
Previously, events wouldn't properly bubble up through parents. This
commit fixes that, though it is admittedly still a bit messy.

Without this commit, it could happen that a group was a made without
children, and if the group later had children added (or if its children
had children added), events would not properly be seen by the viewmodel.